Moved CImgRender to its own folder
