+++ adagucserverEC/CImgRenderers/CImgRenderPoints.h
