+++ adagucserverEC/COpenDAPHandler.h
@@ -8,7 +8,6 @@
 #include "CImgWarpNearestRGBA.h"
 #include "CImgWarpBilinear.h"
 #include "CImgWarpBoolean.h"
-#include "CImgRenderPoints.h"
 #include "CStyleConfiguration.h"
 #include "CMyCURL.h"
 #include "CXMLParser.h"
